refactor(claude-cli): centralize terminal error conversion
