+++ src/agents/cli-runner/output-error.ts
@@ -1,0 +1,37 @@
+import { formatCliOutputError, type CliOutput } from "../cli-output.js";
+import { classifyFailoverReason } from "../embedded-agent-helpers.js";
+import { FailoverError, resolveFailoverStatus } from "../failover-error.js";
+
+export function createCliOutputFailoverError(params: {
+  output: CliOutput;
+  provider: string;
+  model: string;
+  runId?: string;
+  sessionId?: string;
+  lane?: string;
+}): FailoverError | undefined {
+  if (!params.output.errorText) {
+    return undefined;
+  }
+  const message = formatCliOutputError(params.output, {
+    runId: params.runId,
+    sessionId: params.sessionId,
+  });
+  const reason = classifyFailoverReason(message, { provider: params.provider }) ?? "unknown";
+  const code =
+    params.output.terminalFailure?.reason === "max_turns"
+      ? "cli_max_turns"
+      : reason === "context_overflow"
+        ? "cli_context_overflow"
+        : undefined;
+  return new FailoverError(message, {
+    reason,
+    provider: params.provider,
+    model: params.model,
+    sessionId: params.sessionId,
+    lane: params.lane,
+    status: resolveFailoverStatus(reason),
+    code,
+    rawError: params.output.errorText,
+  });
+}
